[FORMAT] added initial lexer
FORMAT is a custom format for writing in human readable text
configurations for the game engine. The intent is to write it '.scheme'
files and export a binaries that the engine can be built into the
executable but leaves everything exposed. Long term goals are to add an
editor platform that take advantage of 'scene' files in binary or text
depending on context